+++ cedar-wasm/CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `policy_text_to_json`, and `policy_text_from_json`. (#616)
 - Exposed cedar-wasm functionality for authorization and validation: `wasm_is_authorized`
   and `wasm_validate`. (#657)
+- Exposed types through `tsify` for `ValidateCall` and the schema. (#692)